Problem: bigint is required by other crates Solution: put bigint into a seperate crate
