+++ bigint/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "bigint"
+version = "0.1.0"
+authors = ["Wei Tang <hi@that.world>"]
+
+[dependencies]
+rlp = { path = '../rlp' }
